Shopping item notifications! - Tapping a shopping item will create a notification, reminding you to buy it.
The background service has returned. This is in preparation for something cool later on. Also added some classes to handle starting it up and binding it to the application.
Renamed the old 'BackgroundService' class to WebHandler, to better represent what it does, as well as to remove conflicts with the actual service class.