Removing Emulator Tests for now
They are not working correctly and I do not know enough to correctly traverse the entire UI in a test with mocked out implementations to prevent API calls to a server. Going to start writing some unit tests for core functionality and begin to rewrite sections of the app for easy extension and modification... as well as easier mocking.